cache Net::HTTP object and force keep-alive
fetching a whole ical file can be hundreds of separate requests to
get every event.  doing this over ssl has lots of connection setup
overhead, but since we're making all of the requests to the same
host anyway and in rapid succession, re-use the connection.  if the
connection closes, Net::HTTP will re-open it for us anyway.